+++ src/protocol/types/game/action/fight/mod.rs
@@ -2,4 +2,4 @@ use std::io::{Read, Write};
 use std::io::Result;
 use protocol::*;
  use variants::AbstractFightDispellableEffectVariant;
-impl_type!(FightDispellableEffectExtendedInformations, 208, action_id| VarShort, source_id| i32, effect| AbstractFightDispellableEffectVariant);
+impl_type!(FightDispellableEffectExtendedInformations, 208, action_id| VarShort, source_id| f64, effect| AbstractFightDispellableEffectVariant);
